fix: rewrite mapping and matching routines
pyannote.core 3.x changes the return type of Annotation.__mul__ (from xarray.DataArray to np.ndarray) because it was found that recent changes in xarray led to inconsistency in diarization metrics. "xarray" is still used in IdentificationErrorAnalysis.matrix but is meant to be removed completely in the future.

Also, "scipy" now provides an implementation of the Hungarian algorithm (supposedly faster than that of "munkres") so we switch to a recent version of "scipy".

Therefore, the code for "HungarianMapper" and "GreedyMapper" has been rewritten to rely on those two things. This also has the side effect of removing both "networkx" and "munkres" dependency.